% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{clustApply}
\alias{clustApply}
\title{Apply a function to each cluster from an hclust object, at a given height.}
\usage{
clustApply(hclustObject, height, FUN, ...)
}
\arguments{
\item{hclustObject}{An hclust object, the result of calling clusterUsers}

\item{height}{Numeric, indicates the height at which to cut the dendogram
and take groups.}

\item{FUN}{A function, one of whose arguments is a set of user_ids.}

\item{...}{Additional arguments to pass to FUN}
}
\value{
A named list of objects, of type value(FUN), one for each cluster in
hclustObject, when cut at height 'height'.
}
\description{
Apply a function to each cluster from an hclust object, at a given height.
}

