% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{clusterUsers}
\alias{clusterUsers}
\title{Cluster users according to their platform action distributions.}
\usage{
clusterUsers(paDistData, distParams = NULL, hclustParams = list(method =
  "ward.D"))
}
\arguments{
\item{paDistData}{A data frame, the result of calling calculatePADist}

\item{distParams}{Named list of additional parameteres to pass to dist}

\item{hclustParams}{Named list of additional parameters to pass to hclust}
}
\value{
An object of type hclust whose labels correspond to the user ids
that appear in paDistDataWide
}
\description{
Convenience function that chains calls to dist and hclust. The hclust method
defaults to ward.D since that's produced the cleanest clusterings in early
exploratory analyses.
}

