% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_network.r
\name{combineClusterRetentionConfounders}
\alias{combineClusterRetentionConfounders}
\title{Combine cluster, individual retention, and confounding variables into a
single data frame.}
\usage{
combineClusterRetentionConfounders(userClust, individualRetention,
  userConfounders, ...)
}
\arguments{
\item{userClust}{An hclust object representing the platform action
distribution clusters for all users.}

\item{individualRetention}{A data.frame, the result of calling
calculateIndividualRetention}

\item{userConfounders}{A data.frame, the result of calling getConfounders}

\item{...}{Additional arguments to pass to clustApply}
}
\value{
A data frame of the form (user_id, week, confounder1, ... confounder
N, active)
}
\description{
Combine cluster, individual retention, and confounding variables into a
single data frame.
}

