% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confound.r
\name{getConfounders}
\alias{getConfounders}
\title{Get a data frame matching each user to the values of several confounding
variables.}
\usage{
getConfounders(users = NULL, queryList, wide = F,
  rundate = as.numeric(gsub(pattern = "-", replacement = "", x = Sys.Date())),
  con = redshift_connection$con)
}
\arguments{
\item{users}{Numeric, a vector of user ids to include in the calculation.}

\item{queryList}{A list of queries to run in the calculation. Each query
calculates one or more confounding variables, and returns the results in a
data.frame.}

\item{wide}{Should results be returned in wide format (T) or long format
(F)?}

\item{rundate}{A dateid of the form yyyymmdd (numeric). All dates after the
rundate will be filtered out.}

\item{con}{The database connection to run the queryList against.}
}
\value{
A data.frame of the form (user_id, variable, value).
}
\description{
Get a data frame matching each user to the values of several confounding
variables.
}

