% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_confounder_belongs_to_cohort_sub}
\alias{query_confounder_belongs_to_cohort_sub}
\title{A string containing the query to calculate whether a user belongs to a
cohort.}
\format{A length-one character vector.}
\usage{
query_confounder_belongs_to_cohort_sub
}
\description{
A string containing the query to calculate whether a user belongs to a
cohort.
}
\keyword{datasets}

