% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_confounder_first_champ_sub}
\alias{query_confounder_first_champ_sub}
\title{A string containing the query to calculate a users' first champion.}
\format{A length-one character vector.}
\usage{
query_confounder_first_champ_sub
}
\description{
A string containing the query to calculate a users' first champion.
}
\keyword{datasets}

