% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_confounder_FL_REVEAL_sub}
\alias{query_confounder_FL_REVEAL_sub}
\title{A string containing the query to calculate whether a user belongs to
FamilyLife or REVEAL for church.}
\format{A length-one character vector.}
\usage{
query_confounder_FL_REVEAL_sub
}
\description{
A string containing the query to calculate whether a user belongs to
FamilyLife or REVEAL for church.
}
\keyword{datasets}

