% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_confounder_oneD7_sub}
\alias{query_confounder_oneD7_sub}
\title{A string containing the query to calculate whether or not a user is a 1D7.}
\format{A length-one character vector.}
\usage{
query_confounder_oneD7_sub
}
\description{
A string containing the query to calculate whether or not a user is a 1D7.
}
\keyword{datasets}

