% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_confounder_use_case}
\alias{query_confounder_use_case}
\title{A string containing the query to calculate the variable 'use_case'}
\format{A length-one character vector.}
\usage{
query_confounder_use_case
}
\description{
A string containing the query to calculate the variable 'use_case'
}
\keyword{datasets}

