% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_individual_retention_sub}
\alias{query_individual_retention_sub}
\title{A string containing the query to calculate the weeks (relative to signup
date) that each user was active.}
\format{A length-one character vector.}
\usage{
query_individual_retention_sub
}
\description{
A string containing the query to calculate the weeks (relative to signup
date) that each user was active.
}
\keyword{datasets}

