% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_pa_dist}
\alias{query_pa_dist}
\title{A string containing the platform action distribution query}
\format{A length-one character vector.}
\usage{
query_pa_dist
}
\description{
A string containing the platform action distribution query
}
\keyword{datasets}

