% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_pa_dist_sub}
\alias{query_pa_dist_sub}
\title{A string containing the platform action distribution query with a placeholder for
subsetting on users and a relative time frame.}
\format{A length-one character vector.}
\usage{
query_pa_dist_sub
}
\description{
A string containing the platform action distribution query with a placeholder for
subsetting on users and a relative time frame.
}
\keyword{datasets}

