% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_retention}
\alias{query_retention}
\title{A string containing the user retention query.}
\format{A length-one character vector.}
\usage{
query_retention
}
\description{
A string containing the user retention query.
}
\keyword{datasets}

