% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_retention_sub}
\alias{query_retention_sub}
\title{A string containing the user retention query with a placeholder for
subsetting.}
\format{A length-one character vector.}
\usage{
query_retention_sub
}
\description{
A string containing the user retention query with a placeholder for
subsetting.
}
\keyword{datasets}

