% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{spreadPADistData}
\alias{spreadPADistData}
\title{Spread the platform action distribution data}
\usage{
spreadPADistData(paDistData)
}
\arguments{
\item{paDistData}{A data frame, the result of calling calculatePADist}
}
\description{
Spreads the platform action ditribution data and adds variable columns if
any platform action groups are missing. This function gets called by
clusterUsers so there will typically be no reason to call it explicitly in
your workflow.
}

