% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{squashPADistList}
\alias{squashPADistList}
\title{Convert an aggPADistList to a single data frame that's ready to be
displayed.}
\usage{
squashPADistList(aggPADistList)
}
\arguments{
\item{aggPADistList}{The result of calling
clustApply(FUN = function(u){
      dplyr::select(calculatePADist(u, agg = T)
                    , flash_report_category, pct_platform_actions)
    }
)
on a userClust object.}
}
\value{
A data frame, one row for each cluster, showing that cluster's
aggregate platform action distribution.
}
\description{
Convert an aggPADistList to a single data frame that's ready to be
displayed.
}

