% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{squashRetentionList}
\alias{squashRetentionList}
\title{Convert a retentionList to a single data frame that's ready for plotting.}
\usage{
squashRetentionList(retentionList)
}
\arguments{
\item{retentionList}{The result of calling
clustApply(FUN=calculateWeeklyRetention) on a userClust object.}
}
\value{
A data.frame of the form  (cluster, relative_session_week,
pct_active)
}
\description{
Convert a retentionList to a single data frame that's ready for plotting.
}

