% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{squashRetentionList}
\alias{squashRetentionList}
\title{Convert a retentionList to a single data frame that's ready for plotting.}
\usage{
squashRetentionList(retentionList)
}
\arguments{
\item{retentionList}{The result of calling
clustApply(FUN=calculateWeeklyRetention) on a userClust object.}
}
\value{
A data.frame of the form  (cluster, <additional variables>
, relative_session_week, pct_active). Here, <additional variables>
represents the variables which appear in the extraGroupings parameter of the
clustApply call.
}
\description{
Convert a retentionList to a single data frame that's ready for plotting.
}

